+++ build.gradle
@@ -5,7 +5,7 @@ plugins {
     id "com.diffplug.spotless" version "6.25.0"
 }
 
-group = 'com.example'
+group = 'com.ureca'
 version = '0.0.1-SNAPSHOT'
 
 spotless {
