+++ src/main/java/com/ureca/DaengGuuBackendApplication.java
@@ -1,4 +1,4 @@
-package com.example.daengguubackend;
+package com.ureca;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
