+++ src/test/java/com/ureca/DaengGuuBackendApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.daengguubackend;
+package com.ureca;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
